create fodler for each component, group all material ui components, create success and error text
