+++ package.json
@@ -24,7 +24,8 @@
     "react-window": "^1.8.4",
     "react-window-infinite-loader": "^1.0.3",
     "styled-components": "^4.3.2",
-    "typescript": "3.5.2"
+    "typescript": "3.5.2",
+    "uuid": "^3.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -53,6 +54,7 @@
     "@types/react-window": "^1.8.0",
     "@types/react-window-infinite-loader": "^1.0.0",
     "@types/styled-components": "^4.1.16",
+    "@types/uuid": "^3.4.5",
     "tslint": "^5.18.0",
     "tslint-react": "^4.0.0"
   }
