+++ src/components/common/ErrorText.tsx
@@ -1,0 +1,7 @@
+import { Typography } from '.'
+import styled from 'styled-components'
+
+export default styled(Typography)`
+  padding-top: 1rem;
+  color: ${props => props.theme.colors.error}
+`
