+++ src/messages.ts
@@ -13,4 +13,6 @@ export default {
   errorTransaction: 'Transactions cannot be performed before users\'ll be loaded.',
   transferSuccess: 'Transfer finished with success!',
   appTitle: 'ETH Explorer',
+  close: 'Close',
+  zeroTransactions: 'User don\'t have any transactions.',
 }
